+++ CHANGELOG.md
@@ -20,6 +20,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### New features
 
+* The 'how to resolve conflicts' hint that is shown when conflicts appear can
+  be hidden by setting `hints.resolving-conflicts = false`.
+
 ### Fixed bugs
 
 ## [0.27.0] - 2025-03-05